Move utilities to their own file
